Move updateRelperms() to the problem class
Moves the updateRelperms() method in BlackOilInstensiveQuantities to the
Problem class. This is a more natural place for this method and it avoids
including EclMaterialManager into the  BlackOilInstensiveQuantities. The
DirectionalMobility struct is moved to a separate file such that it can be
include from both the Problem files and the BlackOilInstensiveQuanMove
updateRelperms() to the problem class
